send_email/requirements.txt
functions-framework==3.*
google-auth-oauthlib
google-api-python-client
google-auth
google-cloud-bigquery




